+++ src/js/helpers/event-bus.js
@@ -63,6 +63,7 @@ export class EventBus extends Vue {
       info: 'info',
       customersUpdateList: 'customers:updateList',
       customersMassIssue: 'customers:massIssue',
+      customerShowBalances: 'customers:showBalances',
       customersHideSelect: 'customers:hideSelect',
       // TODO: should not be here
       businessesSetCurrentBusiness: 'businesses:setCurrentBusiness',
@@ -80,6 +81,7 @@ export class EventBus extends Vue {
       createAssetRequestsUpdateList: 'createAssetRequests:updateList',
       atomicSwapsUpdateList: 'atomicSwaps:updateList',
       updateAccountRole: 'updateAccountRole',
+      promoCodesUpdateList: 'promoCodes:updateList',
     }
   }